use source generator to simplify HttpClient Injection

+++ src/Snap.Hutao/Snap.Hutao.Win32/Unsafe.cs
@@ -1,6 +1,10 @@
 ﻿using Windows.Win32.UI.WindowsAndMessaging;
 
 namespace Snap.Hutao.Win32;
+
+/// <summary>
+/// 包装不安全的代码
+/// </summary>
 public class Unsafe
 {
     /// <summary>
@@ -9,7 +13,7 @@ public class Unsafe
     /// <param name="lParam">lParam</param>
     /// <param name="minWidth">最小宽度</param>
     /// <param name="minHeight">最小高度</param>
-    public static unsafe void SetMinTrackSize(nint lParam, float minWidth, float minHeight)
+    public static unsafe void SetMinTrackSize(nint lParam, double minWidth, double minHeight)
     {
         MINMAXINFO* info = (MINMAXINFO*)lParam;
         info->ptMinTrackSize.x = (int)Math.Max(minWidth, info->ptMinTrackSize.x);
